requirements.txt
connexion[aiohttp,swagger-ui] >= 2.6.0,< 3.0; python_version>="3.6"
# connexion requires werkzeug but connexion < 2.4.0 does not install werkzeug
# we must peg werkzeug versions below to fix connexion
# https://github.com/zalando/connexion/pull/1044
swagger-ui-bundle >= 0.0.9
aiohttp_jinja2 >= 1.5.0
aiohttp_cors >= 0.7.0
cassandra-driver>=3.27.0
openapi-schema-validator >= 0.2.3
numpy>=1.22
eth-event>=1.2.3
lz4~=4.3.2
base58~=2.1
cashaddress>=1.0.6, < 1.1
async-lru~=2.0.4
python-dateutil>=2.9.0
tagpack-tool @ git+https://github.com/graphsense/graphsense-tagpack-tool@feature/interface
